// src/app/core/services/workloads/enums/ConfigurationType.enum.ts
export enum ConfigurationType {
  SkipFavoriteConfig = "SkipFavoriteConfig",
  FloatingReservations = "FloatingReservations",
  MapCapabilities = "MapCapabilities",
  LimitBookingInFuture = "LimitBookingInFuture",
  IsSlackEnabled = "IsSlackEnabled",
  IsTeamsEnabled = "IsTeamsEnabled",
  IsBambooHREnabled = "IsBambooHREnabled",
  IsWorkDayEnabled = "IsWorkDayEnabled",
  IsLuccaEnabled = "IsLuccaEnabled",
  IsPayFitEnabled = "IsPayFitEnabled",
  IsCegedimSRHEnabled = "IsCegedimSRHEnabled",
  IsSuccessFactorEnabled = "IsSuccessFactorEnabled",
  IsHalfDaysEnabled = "IsHalfDaysEnabled",
  IsNeoNomadLayerEnabled = "IsNeoNomadLayerEnabled"
}